+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Helm**
 - Allow values override to also be a URL.
+- Add `keyring` param to `.package` method.
 
 ### 1.2.1
 **General**